OSD: allow recovery pausing and deferral

+++ src/osd/PG.h
@@ -1005,9 +1005,9 @@ public:
    * @returns true if any useful work was accomplished; false otherwise
    */
   virtual bool start_recovery_ops(
-    int max, RecoveryCtx *prctx,
+    uint64_t max, RecoveryCtx *prctx,
     ThreadPool::TPHandle &handle,
-    int *ops_begun) = 0;
+    uint64_t *ops_begun) = 0;
 
   void purge_strays();
